endpoint error messages, table cretion funcs and readme

+++ server/src/tagRouter.ts
@@ -5,19 +5,31 @@ const tagRouter = Router();
 
 tagRouter.post("/:id", async (req, res) => {
   const { off, def, target } = req.body;
-  const result = await dao.setNewTags(req.params.id, off, def, target);
-  res.status(200).send(result);
+  if (off && def && target && req.params.id) {
+    const result = await dao.setNewTags(req.params.id, off, def, target);
+    res.status(201).send(result);
+  } else {
+    res.status(400).send();
+  }
 });
 
 tagRouter.put("/:id", async (req, res) => {
   const { off, def, target } = req.body;
-  const result = await dao.updateTag(req.params.id, off, def, target);
-  res.status(200).send(result);
+  if (off && def && target && req.params.id) {
+    const result = await dao.updateTag(req.params.id, off, def, target);
+    res.status(200).send(result);
+  } else {
+    res.status(400).send();
+  }
 });
 
 tagRouter.get("/:id", async (req, res) => {
-  const result = await dao.findTags(req.params.id);
-  res.status(200).send(result.rows);
+  if (req.params.id) {
+    const result = await dao.findTags(req.params.id);
+    res.status(200).send(result.rows);
+  } else {
+    res.status(400).send();
+  }
 });
 
 tagRouter.get("/", async (_req, res) => {
@@ -26,10 +38,12 @@ tagRouter.get("/", async (_req, res) => {
 });
 
 tagRouter.delete("/:id", async (req, res) => {
-  const result = await dao.deleteTag(req.params.id);
-  res.status(200).send(result.rows);
+  if (req.params.id) {
+    const result = await dao.deleteTag(req.params.id);
+    res.status(200).send(result.rows);
+  } else {
+    res.status(400).send("missing params");
+  }
 });
-
-
 
 export default tagRouter;
